+++ R/dataframe.R
@@ -193,6 +193,12 @@ as_data_frame.matrix <- function(x, ...) {
   matrixToDataFrame(x)
 }
 
+#' @export
+#' @rdname as_data_frame
+as_data_frame.tbl_cube <- function(x, ...) {
+  as_data_frame(as.data.frame(x, ..., stringsAsFactors = FALSE))
+}
+
 #' Convert row names to an explicit variable.
 #'
 #' @param df Input data frame with rownames.